bookmark-backend/requirements.txt
asgiref==3.8.1
Django==5.1.2
sqlparse==0.5.1
asgiref==3.8.1
Django==5.1.2
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
iniconfig==2.0.0
packaging==24.1
pluggy==1.5.0
PyJWT==2.9.0
pytest==8.3.3
pytest-django==4.9.0
pytest-pythonpath==0.7.3
sqlparse==0.5.1
